+++ lib/client.ts
@@ -35,33 +35,69 @@ export default class Client {
     );
   }
 
-  public async pushMessage(
-    to: string,
-    messages: Types.Message | Types.Message[],
-  ): Promise<any> {
-    return this.http.post("/message/push", {
-      messages: toArray(messages),
-      to,
+  private setLineRequestId(response: object, lineRequestId: string): boolean {
+    return Reflect.defineProperty(response, "getLineRequestId", {
+      enumerable: false,
+      value: (): string => {
+        return lineRequestId;
+      },
     });
   }
 
-  public async replyMessage(
+  private async postMessagingAPI(
+    url: string,
+    body?: any,
+  ): Promise<Types.MessageAPIResponseBase> {
+    const res = await this.http.postJson(url, body);
+    // header names are lower-cased
+    // https://nodejs.org/api/http.html#http_message_headers
+    this.setLineRequestId(res.data, res.headers["x-line-request-id"]);
+    return res.data as Types.MessageAPIResponseBase;
+  }
+
+  public pushMessage(
+    to: string,
+    messages: Types.Message | Types.Message[],
+    notificationDisabled: boolean = false,
+  ): Promise<Types.MessageAPIResponseBase> {
+    return this.postMessagingAPI("/message/push", {
+      messages: toArray(messages),
+      to,
+      notificationDisabled,
+    });
+  }
+
+  public replyMessage(
     replyToken: string,
     messages: Types.Message | Types.Message[],
-  ): Promise<any> {
-    return this.http.post("/message/reply", {
+    notificationDisabled: boolean = false,
+  ): Promise<Types.MessageAPIResponseBase> {
+    return this.postMessagingAPI("/message/reply", {
       messages: toArray(messages),
       replyToken,
+      notificationDisabled,
     });
   }
 
   public async multicast(
     to: string[],
     messages: Types.Message | Types.Message[],
-  ): Promise<any> {
-    return this.http.post("/message/multicast", {
+    notificationDisabled: boolean = false,
+  ): Promise<Types.MessageAPIResponseBase> {
+    return this.postMessagingAPI("/message/multicast", {
       messages: toArray(messages),
       to,
+      notificationDisabled,
+    });
+  }
+
+  public async broadcast(
+    messages: Types.Message | Types.Message[],
+    notificationDisabled: boolean = false,
+  ): Promise<any> {
+    return this.http.post("/message/broadcast", {
+      messages: toArray(messages),
+      notificationDisabled,
     });
   }
 
@@ -253,6 +289,33 @@ export default class Client {
   ): Promise<Types.NumberOfMessagesSentResponse> {
     const res = await this.http.get<Types.NumberOfMessagesSentResponse>(
       `/message/delivery/multicast?date=${date}`,
+    );
+    return ensureJSON(res);
+  }
+
+  public async getTargetLimitForAdditionalMessages(): Promise<
+    Types.TargetLimitForAdditionalMessages
+  > {
+    const res = await this.http.get<Types.TargetLimitForAdditionalMessages>(
+      "/message/quota",
+    );
+    return ensureJSON(res);
+  }
+
+  public async getNumberOfMessagesSentThisMonth(): Promise<
+    Types.NumberOfMessagesSentThisMonth
+  > {
+    const res = await this.http.get<Types.NumberOfMessagesSentThisMonth>(
+      "/message/quota/consumption",
+    );
+    return ensureJSON(res);
+  }
+
+  public async getNumberOfSentBroadcastMessages(
+    date: string,
+  ): Promise<Types.NumberOfSentBroadcastMessages> {
+    const res = await this.http.get<Types.NumberOfSentBroadcastMessages>(
+      `/message/delivery/broadcast?date=${date}`,
     );
     return ensureJSON(res);
   }