+++ lib/http.ts
@@ -1,4 +1,5 @@
 import axios, { AxiosInstance, AxiosError } from "axios";
+import { AxiosResponse } from "axios";
 import { Readable } from "stream";
 import { HTTPError, ReadError, RequestError } from "./exceptions";
 import * as fileType from "file-type";
@@ -36,10 +37,14 @@ export default class HTTPClient {
     return res.data as Readable;
   }
 
-  public async post<T>(url: string, body?: any): Promise<T> {
-    const res = await this.instance.post(url, body, {
+  public postJson(url: string, body?: any): Promise<AxiosResponse> {
+    return this.instance.post(url, body, {
       headers: { "Content-Type": "application/json" },
     });
+  }
+
+  public async post<T>(url: string, body?: any): Promise<T> {
+    const res = await this.postJson(url, body);
     return res.data;
   }